:bug: : 방안에 유저가 존재할 경우 방장이 방삭제시 예외처리

+++ src/room/room.service.ts
@@ -159,6 +159,8 @@ export class RoomService {
     });
     if (!findRoom) throw RoomException.roomNotFound();
 
+    //방안에 유저가 있는 경우 에러띄움
+    if (findRoom.count) throw RoomException.roomUserexists();
     if (userId != findRoom.ownerId) throw UserException.userUnauthorized();
     const deleteResult = await this.prisma.room.delete({
       where: { uuid: uuid },
